Separate provider embedding response parsing
Move OpenAI and Google embedding response parsing helpers out of runtime-loader.ts so the main runtime module keeps shrinking around request/response orchestration. The embedding runtime factories still call the same helper names, now imported from a focused module.

Constraint: Behavior-preserving refactor only; no runtime API changes or new dependencies
Rejected: Split embedding runtime factories too | would mix helper extraction with factory relocation and widen the review surface
Rejected: Add direct unit tests for the private parsing helpers | existing runtime-loader embedding runtime tests lock behavior through the public runtime surface
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep provider embedding response validation messages stable unless callers/tests are updated deliberately
Tested: VF_DISABLE_LRU_INTERVAL=1 deno test --no-check --allow-all src/provider/runtime-loader.test.ts src/provider/runtime-loader/provider-request-init.test.ts; deno fmt --check src/provider/runtime-loader.ts src/provider/runtime-loader/provider-embedding-responses.ts; DENO_NO_PACKAGE_JSON=1 deno lint src/provider/runtime-loader.ts src/provider/runtime-loader/provider-embedding-responses.ts; deno check src/provider/runtime-loader.ts src/provider/runtime-loader/provider-embedding-responses.ts; deno task typecheck
Not-tested: Full deno task test suite before push; pre-push hook is expected to run it
